python/requirements.txt
apachelogs==0.6.0
cassandra-driver==3.25.0
tqdm==4.64.1